refactor: change .mts to ts
