+++ tests/medicationRoutes.test.ts
@@ -1,9 +1,9 @@
 // tests/medicationRoutes.test.mts
 import request from "supertest";
-import app from "@/app.mts";
-import { AppDataSource } from "@/data-source.mts";
-import { User } from "@/models/User.mts";
-import { Medication } from "@/models/Medication.mts";
+import app from "@/app.ts";
+import { AppDataSource } from "@/data-source.ts";
+import { User } from "@/models/User.ts";
+import { Medication } from "@/models/Medication.ts";
 
 describe("Medication Routes", () => {
   let token = "";
