+++ tests/middlewares.test.ts
@@ -1,13 +1,13 @@
 /// tests/middlewares.test.mts ///
 
 import request from "supertest";
-import app from "@/app.mts";
+import app from "@/app.ts";
 import jwt, { JwtPayload } from "jsonwebtoken";
-import { AppDataSource } from "@/data-source.mts";
-import { User } from "@/models/User.mts";
-import { Medication } from "@/models/Medication.mts";
-import { POINTS_CONFIG } from "@/middlewares/pointsMiddleware.mts";
-import config from "../src/config.mts";
+import { AppDataSource } from "@/data-source.ts";
+import { User } from "@/models/User.ts";
+import { Medication } from "@/models/Medication.ts";
+import { POINTS_CONFIG } from "@/middlewares/pointsMiddleware.ts";
+import config from "../src/config.ts";
 
 describe("Middleware Tests", () => {
   let token: string;
