+++ tsconfig.json
@@ -29,7 +29,7 @@
   "include": [
     "next-env.d.mts",
     ".next/types/**/*.ts",
-    "**/*.mts",
+    "**/*.ts",
     "**/*.tsx",
     "components/button/icon-button/RoundButton.ts"
   ],
